# coding-openapi.yaml
openapi: "3.1.0"
info:
  title: "GPT-API Coding Agent API"
  version: "1.0.0"
  description: "Narrow Custom GPT Coding Agent API with strict state-machine contracts."
paths:
...
  /repo/callgraph:
    post:
      summary: "Repo Callgraph"
      operationId: "repo_callgraph_repo_callgraph_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/references:
    post:
      summary: "Repo References"
      operationId: "repo_references_repo_references_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoReferencesRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/symbol-references:
    post:
      summary: "Repo Symbol References"
      operationId: "repo_symbol_references_repo_symbol_references_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoSymbolReferencesRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/route-map:
    post:
      summary: "Repo Route Map"
      operationId: "repo_route_map_repo_route_map_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/changed-context:
    post:
      summary: "Repo Changed Context"
      operationId: "repo_changed_context_repo_changed_context_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoChangedContextRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/recent-history-context:
    post:
      summary: "Repo Recent History Context"
      operationId: "repo_recent_history_context_repo_recent_history_context_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoRecentHistoryRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
  /policy/evaluate-action-deep:
    post:
      summary: "Policy Evaluate Action Deep"
      operationId: "policy_evaluate_action_deep_policy_evaluate_action_deep_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyEvaluateDeepRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /agent/coding-task:
    post:
      summary: "Coding Task"
      operationId: "coding_task_agent_coding_task_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodingTaskRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
  /tasks/status-summary:
    post:
      summary: "Task Status Summary"
      operationId: "task_status_summary_tasks_status_summary_post"
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

      security:
        -
          ApiKeyAuth:
  /tasks/gc:
    post:
      summary: "Task Gc"
      operationId: "task_gc_tasks_gc_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskGcRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /tasks/lock-ttl:
    post:
      summary: "Task Lock Ttl"
      operationId: "task_lock_ttl_tasks_lock_ttl_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskLockTtlRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /tasks/artifact-index:
    post:
      summary: "Task Artifact Index"
      operationId: "task_artifact_index_tasks_artifact_index_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskIdRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
  /github/pr/update-body:
    post:
      summary: "Pr Update Body"
      operationId: "pr_update_body_github_pr_update_body_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPrUpdateBodyRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/review-comments:
    post:
      summary: "Pr Review Comments"
      operationId: "pr_review_comments_github_pr_review_comments_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubReviewCommentsRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/checks/logs:
    post:
      summary: "Checks Logs"
      operationId: "checks_logs_github_checks_logs_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubChecksLogsRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/branch/push:
    post:
      summary: "Branch Push"
      operationId: "branch_push_github_branch_push_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubBranchPushRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
  /diagnostics/parse:
    post:
      summary: "Diagnostics Parse"
      operationId: "diagnostics_parse_diagnostics_parse_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticsParseRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
  /diagnostics/triage:
    post:
      summary: "Diagnostics Triage"
      operationId: "diagnostics_triage_diagnostics_triage_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticsTriageRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /env/discover:
    post:
      summary: "Env Discover"
      operationId: "env_discover_env_discover_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /env/doctor:
    post:
      summary: "Env Doctor"
      operationId: "env_doctor_env_doctor_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /env/prepare-dry-run:
    post:
      summary: "Env Prepare Dry Run"
      operationId: "env_prepare_dry_run_env_prepare_dry_run_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
components:
  schemas:
...
    CodingTaskRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        task:
          type: "string"
          title: "Task"
        mode:
          type: "string"
          title: "Mode"
          default: "plan_apply_verify"
        workspace_strategy:
          type: "string"
          title: "Workspace Strategy"
          default: "git_worktree"
        max_iterations:
          type: "integer"
          title: "Max Iterations"
          default: 5
        approval_policy:
          type: "string"
          title: "Approval Policy"
          default: "safe_auto"
        create_pr:
          type: "boolean"
          title: "Create Pr"
          default: false
        base_branch:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Base Branch"
        task_id:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Task Id"
      type: "object"
      required:
        - "repo_path"
        - "task"
      title: "CodingTaskRequest"
...
    DiagnosticsParseRequest:
      properties:
        tool:
          type: "string"
          title: "Tool"
        stdout:
          type: "string"
          title: "Stdout"
          default: ""
        stderr:
          type: "string"
          title: "Stderr"
          default: ""
      type: "object"
      required:
        - "tool"
      title: "DiagnosticsParseRequest"
...
    DiagnosticsTriageRequest:
      properties:
        diagnostics:
          items:
            additionalProperties: true
            type: "object"
          type: "array"
          title: "Diagnostics"
        task:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Task"
        max_files:
          type: "integer"
          title: "Max Files"
          default: 20
      type: "object"
      required:
        - "diagnostics"
      title: "DiagnosticsTriageRequest"
...
    EnvRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
      type: "object"
      required:
        - "workspace_path"
      title: "EnvRequest"
...
    GithubBranchPushRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        remote:
          type: "string"
          title: "Remote"
          default: "origin"
        branch:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Branch"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
      title: "GithubBranchPushRequest"
...
    GithubChecksLogsRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        ref:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Ref"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
      title: "GithubChecksLogsRequest"
...
    GithubPrUpdateBodyRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        pr:
          type: "string"
          title: "Pr"
        body:
          type: "string"
          title: "Body"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
        - "pr"
        - "body"
      title: "GithubPrUpdateBodyRequest"
    GithubReviewCommentsRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        pr:
          type: "string"
          title: "Pr"
      type: "object"
      required:
        - "workspace_path"
        - "pr"
      title: "GithubReviewCommentsRequest"
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    PolicyEvaluateDeepRequest:
      properties:
        action:
          type: "string"
          title: "Action"
        workspace_path:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Workspace Path"
        changed_files:
          anyOf:
            -
              items:
                type: "string"
              type: "array"
            -
              type: "null"
          title: "Changed Files"
        tests_passed:
          anyOf:
            -
              type: "boolean"
            -
              type: "null"
          title: "Tests Passed"
        quality_passed:
          anyOf:
            -
              type: "boolean"
            -
              type: "null"
          title: "Quality Passed"
        user_approved_network_write:
          type: "boolean"
          title: "User Approved Network Write"
          default: false
        user_approved_sensitive:
          type: "boolean"
          title: "User Approved Sensitive"
          default: false
        user_approved_large_diff:
          type: "boolean"
          title: "User Approved Large Diff"
          default: false
        user_approved_deletions:
          type: "boolean"
          title: "User Approved Deletions"
          default: false
        diff_line_count:
          type: "integer"
          title: "Diff Line Count"
          default: 0
      type: "object"
      required:
        - "action"
      title: "PolicyEvaluateDeepRequest"
...
    RepoChangedContextRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        base_ref:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Base Ref"
      type: "object"
      required:
        - "repo_path"
      title: "RepoChangedContextRequest"
    RepoOverviewRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        max_depth:
          type: "integer"
          title: "Max Depth"
          default: 4
      type: "object"
      required:
        - "repo_path"
      title: "RepoOverviewRequest"
...
    RepoRecentHistoryRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        query:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Query"
        max_commits:
          type: "integer"
          title: "Max Commits"
          default: 20
      type: "object"
      required:
        - "repo_path"
      title: "RepoRecentHistoryRequest"
    RepoReferencesRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        symbol:
          type: "string"
          title: "Symbol"
        max_results:
          type: "integer"
          title: "Max Results"
          default: 100
      type: "object"
      required:
        - "repo_path"
        - "symbol"
      title: "RepoReferencesRequest"
...
    RepoSymbolReferencesRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        symbols:
          items:
            type: "string"
          type: "array"
          title: "Symbols"
        max_results_per_symbol:
          type: "integer"
          title: "Max Results Per Symbol"
          default: 50
      type: "object"
      required:
        - "repo_path"
        - "symbols"
      title: "RepoSymbolReferencesRequest"
...
    TaskGcRequest:
      properties:
        max_age_ms:
          type: "integer"
          title: "Max Age Ms"
          default: 604800000
        statuses:
          anyOf:
            -
              items:
                type: "string"
              type: "array"
            -
              type: "null"
          title: "Statuses"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      title: "TaskGcRequest"
    TaskIdRequest:
      properties:
        task_id:
          type: "string"
          title: "Task Id"
      type: "object"
      required:
        - "task_id"
      title: "TaskIdRequest"
...
    TaskLockTtlRequest:
      properties:
        task_id:
          type: "string"
          title: "Task Id"
        owner:
          type: "string"
          title: "Owner"
          default: "coding-gpt"
        ttl_ms:
          type: "integer"
          title: "Ttl Ms"
          default: 1800000
      type: "object"
      required:
        - "task_id"
      title: "TaskLockTtlRequest"
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
...
  securitySchemes:
    ApiKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-api-key"
servers:
  -
    url: "https://gpt-api.ngrok.app"
security:
  -
    ApiKeyAuth: